# docs/admin/admin_swagger.yaml
basePath: /admin-api
definitions:
...
    handler.adminResponse:
        properties:
            code:
                type: integer
            data: {}
            msg:
                type: string
        type: object
...
    handler.currentUserDocResponse:
        properties:
            avatar:
                type: string
            enabled:
                type: boolean
            id:
                type: integer
            name:
                type: string
            permissions:
                items:
                    type: string
                type: array
            username:
                type: string
        type: object
...
    handler.loginDocResponse:
        properties:
            token:
                type: string
        type: object
    handler.loginRequest:
        properties:
            captchaCode:
                type: string
            captchaId:
                type: string
            password:
                type: string
            username:
                type: string
        required:
            - captchaCode
            - captchaId
            - password
            - username
        type: object
...
    handler.uploadResponse:
        properties:
            url:
                type: string
        type: object
...
    service.ChangePasswordRequest:
        properties:
            newPassword:
                minLength: 8
                type: string
            oldPassword:
                type: string
        required:
            - newPassword
            - oldPassword
        type: object
    service.UpdateProfileRequest:
        properties:
            avatar:
                type: string
            name:
                type: string
        type: object
...
host: localhost:8080
info:
    contact:
        name: API Support
        url: https://github.com/slowlyo/bico-admin
    description: 后台管理模块 API 文档
    license:
        name: MIT
        url: https://opensource.org/licenses/MIT
    termsOfService: https://github.com/slowlyo/bico-admin
    title: Bico Admin Admin API
    version: "1.0"
paths:
...
    /auth/avatar:
        post:
            consumes:
                - multipart/form-data
            description: 上传当前用户头像文件
            parameters:
                - description: 头像文件
                  in: formData
                  name: avatar
                  required: true
                  type: file
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        allOf:
                            - $ref: '#/definitions/handler.adminResponse'
                            - properties:
                                data:
                                    $ref: '#/definitions/handler.uploadResponse'
                              type: object
            security:
                - BearerAuth: []
            summary: 上传头像
            tags:
                - 认证
    /auth/current-user:
        get:
            description: 获取当前登录用户资料和权限
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        allOf:
                            - $ref: '#/definitions/handler.adminResponse'
                            - properties:
                                data:
                                    $ref: '#/definitions/handler.currentUserDocResponse'
                              type: object
            security:
                - BearerAuth: []
            summary: 获取当前用户
            tags:
                - 认证
    /auth/login:
        post:
            consumes:
                - application/json
            description: 使用账号、密码和验证码换取登录 token
            parameters:
                - description: 登录参数
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/handler.loginRequest'
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        allOf:
                            - $ref: '#/definitions/handler.adminResponse'
                            - properties:
                                data:
                                    $ref: '#/definitions/handler.loginDocResponse'
                              type: object
            summary: 登录
            tags:
                - 认证
    /auth/logout:
        post:
            description: 将当前 token 加入黑名单
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.adminResponse'
            security:
                - BearerAuth: []
            summary: 退出登录
            tags:
                - 认证
    /auth/password:
        put:
            consumes:
                - application/json
            description: 修改当前登录用户密码
            parameters:
                - description: 密码参数
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/service.ChangePasswordRequest'
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.adminResponse'
            security:
                - BearerAuth: []
            summary: 修改密码
            tags:
                - 认证
    /auth/profile:
        put:
            consumes:
                - application/json
            description: 更新当前登录用户的名称和头像
            parameters:
                - description: 个人资料
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/service.UpdateProfileRequest'
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        allOf:
                            - $ref: '#/definitions/handler.adminResponse'
                            - properties:
                                data:
                                    $ref: '#/definitions/handler.currentUserDocResponse'
                              type: object
            security:
                - BearerAuth: []
            summary: 更新个人资料
            tags:
                - 认证
...
securityDefinitions:
    BearerAuth:
        description: 'JWT 认证，格式: Bearer {token}'
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"